Added a loading module, added Dungeon class to DEM_main
Probably should have done two commits here. Whatever, I'm still learning. I created a VERY barebones module that will serve to load and unload (and create) dungeon environments for the DEM, as well as adding a Dungeon class to DEM_main. Honestly though, I might get rid of this and put it in yet another new module. Frankly this commit is mostly to serve as a reminder that this was, in the past, an implementation I thought about. And isn't that what makes git so wonderful?